feat(mydsahboard): implement my dashboard. refs #24
